src/api/requirements.txt
flask
paho-mqtt
coloredlogs